refactor: modularize plot graph
